[#207] formview: organize component as trees
Previously the components where hard to navigate. Let's use a more
classical approach to sort them out.
