--- src/brokers/add-broker/components/AddBrokerForm/index.ts
@@ -1,2 +1,0 @@
-export * from './YamlEditorView';
-export * from './FormView';
