--- src/brokers/add-broker/components/index.ts
@@ -1,2 +1,0 @@
-export * from './AddBrokerForm';
-export * from './EditorToggle';
